minimal-bootstrap.tinycc-{bootstrappable,mes}: Support x86_64-linux
Because common.nix is shared across these packages, they are upgraded together.
This patch applies various x86_64-specific bugfixes to tinycc-bootstrappable
and upgrades tinycc-mes to the latest "mob" version.

The build scripts are updated to support 64-bit targets. The runtime library
"libtcc1.c" is changed to include va_list.c for tinycc-bootstrappable, which is
required to support variadic args on x86_64. On the other hand, we add alloca
support to libtcc1.c on tinycc-mes.