minimal-bootstrap.musl: Support x86_64-linux
We now build tinycc-musl through an intermediate stage.
BEFORE:
tinycc-mes (links to mes-libc) -> musl -> tinycc-musl (links to musl)
NOW:
tinycc-mes (links to mes-libc) -> musl-intermediate
  -> tinycc-musl-intermediate -> musl -> tinycc-musl

This fixes some issues with handling floating-point numbers:
certain float routines are not implemented are not implemented in mes-libc,
which would cause a miscompiled musl in the old approach.

Certain unnecessary musl patches are dropped. It is especially important to
remove the weak symbol patch, which in fact breaks the build.

tinycc-musl is upgraded to the latest "mob" version, with a couple bugfixes.