openscad: do not remove CLI for Darwin
In its prior form, OpenSCAD would be made available via a typical macOS `.app`
bundle.  OpenSCAD also provides a useful command line interface that is helpful
for automation and other scripted activities.  The `openscad` binary used by the
`.app` bundle works as the CLI.  These changes keep the original `openscad`
binary and add it to `bin` where it will automatically wind up on the `PATH` for
consumers of this package.