stdenv: fix inputDerivation in case of __structuredAttrs
Setting allowedReferences to null seems to only work as a fluke. It
doesn't work with outputChecks, and I couldn't get it to work at all
when declaring my own derivation manually (I'm honestly still unsure why
it works at all as-is in inputDerivation)

In any case, Rather than setting allowedReferences etc to values that
mimic the default behaviour, we can remove those attributes to get the
default behaviour.