lib.attrsets.recursiveUpdateUntil: rewrite
We make the assumption that on average, attribute sets end up being more
disjoint than similar. because of this, we now run `intersectAttrs` on
the two attribute sets in every recursive call. If the attribute sets
are nearly the same, this is a performance regression - but that will
account for a minority of cases. to see if we can skip a `zipAttrsWith`
call and exit early.