nixos-rebuild-ng: improve IPv6 support
These changes improve support for using an IPv6 address for the target
host. This approach is similar to the one found in nix [1].
Previously, nixos-rebuild-ng would pass a host string to `ssh` that
cannot be parsed.

Since the host string is being treated like a URL, we encode the scope
ID symbol (%) as %25. This is required for compatibility with other nix
commands (e.g. nix-copy-closure).

Example:
```
nixos-rebuild \
  --target-host "root@[fe80::9272:6aff:fe15:faf3%25enp4s0]" \
  --flake path:/etc/nixos#xiatian -L \
  switch
```

[1] https://github.com/NixOS/nix/blob/8b955d8/src/libstore/store-reference.cc#L143-L166

Signed-off-by: David Wronek <david.wronek@mainlining.org>