Fix CVS on Darwin
Currently, invocations of CVS (and by extension `fetchcvs`) usually die
immediately on Darwin with

    cvs [checkout aborted]: received abort signal

The root cause is a Homebrew patch that is incompatible with the Debian
patchset.  Currently nixpkgs pulls in both.  This commit fixes the
problem by removing the Homebrew patch.

This is the Homebrew patch:

    --- a/lib/vasnprintf.c  2017-01-01 00:00:00.000000000 -0000
    +++ b/lib/vasnprintf.c  2017-01-01 00:00:00.000000000 -0000
    @@ -558,11 +558,13 @@
              }
            *p = dp->conversion;
     #if USE_SNPRINTF
    +# if ! (defined __APPLE__ && defined __MACH__)
            p[1] = '%';
            p[2] = 'n';
            p[3] = '\0';
    -#else
    +# else
            p[1] = '\0';
    +# endif
     #endif

            /* Construct the arguments for calling snprintf or sprintf.  */

A close inspection reveals that the patch misbehaves when
`USE_SNPRINTF == 0`:

 - Without the patch, the line `p[1] = '\0';` would run.
 - With the patch, the entire block is skipped, and no 0 byte is
   written anywhere!

The patch normally works because on Darwin `USE_SNPRINTF == 1`.  But
the Debian patchset hard-codes `USE_SNPRINTF == 0`:

    +#if 0
    +/* disabled for security reasons, to avoid having %n in writable memory */
     # define USE_SNPRINTF (HAVE_DECL__SNPRINTF || HAVE_SNPRINTF)
    +#else
    +# define USE_SNPRINTF 0
    +#endif

All this is to say that we don't need both patches.  The Debian
patchset alone suffices.