python3Packages.textual: Use explicit tree-sitter language dependencies
tree-sitter-languages is deprecated and broken on the latest version of
tree-sitter. Thus, textual doesn't support syntax highlighting of any
languages right now. tree-sitter-language-pack is a replacement fork for
tree-sitter-languages, but isn't plug-and-play with textual. Add
dependencies for all the currently Python-packaged tree-sitter
languages, so that at least some languages work with textual.