nixos-rebuild-ng: remove nixosTests
Since the recent changes that makes any PR that rebuilds NixOS tests to
target `staging-nixos`, landing changing in `nixos-rebuild-ng` is
becoming annoying since we need to wait for the longer release cycle of
that branch.

Running `nixosTests` for every change is probably unnecessary because
different from the old `nixos-rebuild`, `nixos-rebuild-ng` includes lots
of unit tests so it is unlikely to break.

So let's remove it for now. We can still manually run nixosTests in a
specific PR with OfBorg if needed.