pipeline: fix playback errors; overhaul build
Adds build fixes based on the flake file build process present in the
upstream repository.

Peertube videos play back normally, but has a "duplication" effect at
certain screen sizes. YouTube videos do not play at all due to an
error reported back by the yt-dlp plugin that the requested format
cannot be found.

A workaround for this is to rely on the external player which
functions as expected with mpv.