ponyc: Correct meta.homepage to correct domain
The main domain for the ponylang project moved from ponylang.org
to ponylang.io in March 2019.

cite: https://github.com/ponylang/ponylang-website/pull/412

The domain is a redirect, which is why it hasn't been noticed
before, but the meta-data in nixpkgs should reflect the domain
owned by the Pony Project.