renode: dotnet sdk 6 -> 9
When we pass the `TargetFrameworks` in `dotnetFlags` we got some weird
unexpected behavior, the `generateAssemblyInfo.sh` wasn't executed,
AssemblyInfo is not generated and we hit
[this](https://github.com/renode/renode/blob/master/src/Renode/Program.cs#L161)
exception.

But this `generateAssemblyInfo` script only replaces version and commit
rev in a `.template` file to generate a `.cs` file. We can do this step
via Nix and everything works correctly.

Signed-off-by: Samuel Silva <samuel@ossystems.com.br>