freecad: fix filebrowser crash on Gnome 49
It seems the gtk3 schema paths aren't present in XDG_DATA_DIRS on
Gnome 49 and thus attempts to open a gtk3 filebrowser will fail,
causing the application to crash. This commit adds the gtk3
gsettings schemas to the XDG_DATA_DIRS env var for the app via the
wrapper script.

Fixes #467783