nixos/fail2ban: load xt_multiport kernel module
The default ban action `iptables-multiport` needs to have the
`xt_multiport` kernel module loaded to work.

Normally this wouldn't be a problem
as it would be automatically loaded on demand,
but it's not the case when `security.lockKernelModules` is `true`.

Now we load it on boot so that it works in any case.