gdb: rename misnamed hostCpuOnly to targetCpuOnly
Cross-compiled GDB with `hostCpuOnly` set to true can debug the target
architecture, not the host architecture. For example, on my machine,
`pkgsCross.arm-embedded.buildPackages.gdbHostCpuOnly` is compiled with
`--host=x86_64-unknown-linux-gnu --target=arm-none-eabi`, and is meant
to debug ARM, even if its name would make you think it's for x86-64.