python3Packages.python-dbusmock: backport networkmanager 1.54.2 fix
This is included in the latest release of dbusmock, but we can't
update to that without a compatibility fix for polkit, which would
need to go through staging.  We already have the networkmanager update
(which has security fixes): in master, so for now, backport the patch
without causing a mass rebuild.  We can then revert this on staging.

Fixes: 403d46a327f1 ("networkmanager: 1.54.1 -> 1.54.3")