nixos/home-assistant: do not set server_host by default
Allowing `services.home-assistant.config.http.server_host` to be set
to `null` and making this default.

Setting the value to null leads to it not being set in the by formats.ya
This should not change the default behaviour of home-assistant, as
by default the process binds to all available addresses.

This change avoids creating a warning in home-assistant about the deprecation
of the server_host option and expected removal in 2026.06

An open issue for upstream around the removal can be found here:
https://github.com/home-assistant/core/issues/157961