ir-lv2: remove mainProgram
The mainProgram was set in a treewide effort but does not actually
reflect the main program.

From
https://github.com/NixOS/nixpkgs/blob/master/pkgs/README.md#meta-attributes:

meta.mainProgram must be set to the name of the executable which
facilitates the primary function or purpose of the package.

Packages which are not primarily used for a single executable do not need to set meta.mainProgram.

(cherry picked from commit abbc5dfdae6450bb4b5ee23b6d54c6c7b82b8076)