pypy: remove explicit darwin minimum SDK
PyPy sets an explicit minimum SDK version for darwin that is much older
than what we we default to on nixpkgs (currently 14.0).
Simply removing the explicit flag makes it use our default instead.

This fixes a build failure where PyPy tries to access APIs that are not
available in the macOS SDK it is targetting.
This is technically not relevant for upstream as this is an opt-in
warning which we have recently enabled by default (as error) on
nixpkgs, but they too should bump the minimum target to at least 10.15
from their current target of 10.13.