aspell: stop using `or` identifier unquoted
This is to be deprecated in Lix:
https://gerrit.lix.systems/c/lix/+/4764/6

The Oriya language is spoken by about 37 million people in India
according to Wikipedia, but I am guessing the dictionary is not actually
used that often by nixpkgs users. Either way quoting it in nixpkgs is
totally non-breaking for users of nixpkgs.

(cherry picked from commit 7e51153b24aedc4430e26b81d29e984210daa23f)