linux/common-config: fix for Linux 5.10.y on aarch64
Until 5.15, HYPERV only supports x86, so forcing module broke
configuration of 5.10 on aarch64.  Our kernel configuration script
sets every module it can to "m" anyway, so it was redundant and
breakage-prone to set any value here for pre-6.18.

I've confiremd that HYPERV is still set to "m" on 6.12 after this
change.

Fixes: e9b977bc41dc ("linux/common-config: update for 6.18")