gh-do: init at 0.5.4
gh-do is a GitHub CLI extension that allows you to "do and undo and
notify" with the GitHub CLI. It provides commands to perform actions
on GitHub resources with the ability to easily revert those actions.

Written in Go, it integrates seamlessly with the gh command line tool.

Homepage: https://github.com/k1LoW/gh-do