librewolf-bin: mark with `knownVulnerabilities` due to lack of maintenance in nixpkgs
This security-critical package is currently not maintained well enough
to be kept as is.
We are currently providing a 4 months old version which is not
acceptable for a browser. PR #464467 created by r-ryantm 3 weeks ago was
not handled by anyone.

See also #416711 for previous discussions.

(cherry picked from commit 1c009d547d610979da59d10a93dbad485942c87f)