gscan2pdf: disable a failing test
The test suite currently fails with

```
    t/04_Page.t ................................... 1/12
        Failed test 'undefined'
        at t/04_Page.t line 114.
            got: '72'
        expected: '300'
    Looks like you failed 1 test of 12.
```

A bisect identified the update of ImageMagick from 7.1.2-8 -> 7.1.2-9 as
the cause of this.  A further bisect of ImageMagick identified commit
ImageMagick@f46647c030 (»Correct parsing of the density«) to be the
actual cause.

I'm lacking the expertise to actually fix this, but informed
[upsteam](https://sourceforge.net/p/gscan2pdf/bugs/439/) about the
problem.

For now, disable the test.

(cherry picked from commit 5cc17f29ad6b24608b25a30d75673cdd974f775a)