gtk3: fix pkg-config wayland and zlib dependencies
Build systems that depend on pkg-config for cflags fail to find wayland headers.
The example I ran into is building qemu in a dev shell environment. This
seems to be caused by nixpkgs patching the behavior of pkg-config so that it
does not include private dependencies in the cflags output unless it's a
static build. I also see examples of this fix in telepathy-glib,
ubports-click, libgweather, and cairo.

The fix here is to move the public wayland-client dependency from
"Requires.private" to "Requires" in gtk+-3.0.pc. This way it shows up in
cflags.

gtk3 also includes zlib in Libs in the gdk-3.0.pc file but does not explicitly
list it as a dependency. This fixes that as well.