java-language-server: avoid using jlink
Hardcode the path to the JDK's java executable in the launch scripts.
Have the build scripts entirely skip using jlink.

This package uses jlink to bundle a Java runtime with its installed
files. This runtime is not complete, and it can't compile user programs
that try to use pieces of the JDK that aren't included (see e.g.
georgewfraser/java-language-server#124).

Upstream, attempts have been made to use the user's JDK instead of the
bundled one to avoid this problem (e.g.
georgewfraser/java-language-server#129). These have been reverted since
this package mucks around with the compiler's internals and thus can
crash in the face of version mismatches.

Nixpkgs has a better option: we can "statically link" this package
against the full JDK used to build it. This allows user code to access a
full JDK while also ensuring that this package can always find its JDK.