nixosTests.syncthing-*: Format
The previous commit was supposed to make sure that
nixos/tests/syncthing/many-devices.nix was properly formatted, but I
guess that our rules for formatting things are different between when
that commit was made in the master branch vs now in the release-25.05
branch.

Additionally, the previous commit did not fix the formatting in
nixos/tests/syncthing/relay.nix.

This change fixes the formatting for all Syncthing NixOS tests.

Not-cherry-picked-because: Potentially different autoformatting rules depending on time of formatting and branch