signal-desktop: fix cross compilation
`nativeBuildInputs = [ (pnpm_10.override { inherit nodejs; }) ]`
uses a `nodejs` built for the host platform, but a `pnpm` built for the
build platform. that causes exec errors, but it seems we don't actually
need `pnpm` to be using the same nodejs version that signal-desktop
itself is built against, so just drop the override.