prisma_6: drop `prisma init` in test which now requires internet access
Later versions of Prisma 6 seem to require internet
access on the first run of `prisma init` to fetch
updates to the subcommand.

This fails in the test environment, but since we
are already sketching out the test schema structure
manually, we can drop that in favour of just creating
the prisma/ directory.