intel-graphics-compiler: fix build with gcc 15
LLVM 16, which this release of IGC is based on, fails to build with
GCC 15 due to a missing include of `<cstdint>` and an include of
`<ciso646>` instead of `<version>`. We patch LLVM directly in order to
fix these build failures.

Additionally, a warning under `-Wfree-nonheap-object` is emitted within
LLVM, and causes compile failures due to `-Werror`. In accordance with
IGC disabling warnings where they originate from within LLVM (see
`IGC/common/LLVMWarningsPush.hpp`), we patch in an ignore for this
warning.