nixos/radicle: Use %d specifier for systemd-creds paths
Was looking for a standard for supporting encrypted creds in systemd
and this seems to be the only instance in nixpkgs.

I noticed the comment says you can't use the %d specifier in
BindReadOnlyPaths, but I've done this before and it works. I wonder
if systemd added this capability since this module was written. So
I updated to use it.

My motivation is that I use encrypted creds extensively and they can
be a real pain to retrofit to existing modules, so I plan to add
support in a few places and want to establish a pattern for doing
this.

If anyone wants to chime in with advice for different ways to expose
this as an option please chime in, but I think this is quite a nice
way to do it (splitting on ":" in the secret path).