dart: fix cross compilation
fixes:
- `nix-build -A pkgsCross.aarch64-multiplatform.buildPackages.dart`,
  as well as its tests:
  - `pkgsCross.aarch64-multiplatform.buildPackages.dart.tests.testCompile`
  - `pkgsCross.aarch64-multiplatform.buildPackages.dart.tests.testCreate`

`bintools` is special, and `pkgs/top-level/all-packages.nix` includes
this note above the `bintools-unwrapped` definition:

> Note when > cross compiling these are used not for this stage but
> the *next* stage. That is why we choose using this stage's target
> platform / next stage's host platform.

i.e. the bintools for "this stage" is actually `buildPackages.bintools`.
that's what splicing would give if the dart derivation accessed bintools
via `nativeBuildInputs = [ bintools ]`.

further evidence for this change is that this existing code is querying
an attribute of the `targetPlatform` based on a property of the `hostPlatform`:
```
lib.optionalString (stdenv.hostPlatform.isLinux) ''
   find $out/bin -executable -type f -exec patchelf --set-interpreter ${bintools.dynamicLinker} {} \;
''
```

`bintools.dynamicLinker` is defined in
pkgs/build-support/bintools-wrapper/default.nix:
```
dynamicLinker =
  if sharedLibraryLoader == null then
    ""
  else if targetPlatform.libc == "musl" then
    "${sharedLibraryLoader}/lib/ld-musl-*"
  else if targetPlatform.libc == "uclibc" then
...
```

we're patching binaries just built for the hostPlatform, so
we can shift the `dynamicLinker` to also be for the hostPlatform
by referencing `buildPackages.bintools.dynamicLinker`.