buildDartApplication: remove unused callPackage args
this fixes cross compilation for at least a few dart applications:
- `nix-build -A pkgsCross.aarch64-multiplatform.domine`
- `nix-build -A pkgsCross.aarch64-multiplatform.unused`

in particular, `dartHooks.override { inherit dart; }` causes
splicing issues (it resolves to a version meant to be used in
`buildInputs`, not `nativeBuildInputs`). but the `dart` argument
turned out to be unused anyways, so just remove it (the hooks pick
up `dart` via the environment -- not via callPackage / substitution).