linux/common-config: Enable `ARM64_PMEM`
This feature, while not enabled by default by the Linux defconfig, is
required for usage of mainline kernels in the NVidia Jetson boards.

It appears to have no adverse effects, since its availability is
detected at runtime, so it's fine to have enabled on all aarch64
platforms.

Other distros have this enabled by default as well:
- Debian: https://salsa.debian.org/kernel-team/linux/-/blob/debian/latest/debian/config/arm64/config#L36
- Ubuntu: https://git.launchpad.net/~ubuntu-kernel/ubuntu/+source/linux/+git/noble/tree/debian.master/config/annotations#n1657
- Fedora: https://src.fedoraproject.org/rpms/kernel/blob/rawhide/f/kernel-aarch64-fedora.config#_497

This is also, from what I can tell, the only blocker to being able to
use NixOS on an NVidia Jetson, since they support UEFI just fine, and
the out-of-tree modules can easily be packaged in nixos-hardware.