dev86: 1.0.1 -> 1.0.1-unstable-2025-02-12, fix build for gcc 15
dev86 fails to build with GCC 15 and C23 due to K&R function definitions
and elided function parameters in *numerous* locations. This change
fixes that in two pieces:

First, dev86 is updated to the current latest `master` release. While
it would normally be odd to go from a stable release to a non-stable
one, this is likely acceptable since all changes between the 1.0.1
release and master have been to fix compilation issues; see
https://codeberg.org/jbruchon/dev86/compare/v1.0.1...0332db1ceb238fa7f98603cdf4223a1d839d4b31.
Additionally, the according to the releases page, the v1.0 release was
primarily so that this could be used in another project, and was largely
untested: https://codeberg.org/jbruchon/dev86/releases/tag/v1.0. Because
of this, it should be fine to move to an unstable version.

After this change, the `unproto` component is the only one which has
remaining build issues. Thus, we include a patch to add previously
elided function parameters.