vicinae: make description more style-guide-compliant
Specifically, the description should not start with an article (the/a/an)
and should avoid the use of personal pronouns like "you". Admittedly this
is a minor and subjective change, but we should be enforcing style
requirements more stringently in Nixpkgs anyway.

(cherry picked from commit 656ae068f23a53a00598685af35bca3eae1168a3)