kernel/common-config: enable all thermal governors
This allows more flexibility in how thermal management is handled in any
NixOS system. It is especially useful for embedded devices, which
often rely on a user-space thermal governor.

Previously, only the `BANG_BANG` and `STEP_WISE` governors were enabled
on x86, or `STEP_WISE` and `POWER_ALLOCATOR` on aarch64, which come from
the defconfig. There is however no negative to having all of them
enabled on all architectures, so for consistency this enables all of
them everywhere.

Other distros seem to have all of the thermal governors enabled as well:
- Ubuntu: https://git.launchpad.net/~ubuntu-kernel/ubuntu/+source/linux/+git/noble/tree/debian.master/config/annotations#n13757
- Debian: https://salsa.debian.org/kernel-team/linux/-/blob/debian/latest/debian/config/config#L5116-L5120
- Arch: https://gitlab.archlinux.org/archlinux/packaging/packages/linux/-/blob/main/config?ref_type=heads#L5788-L5792