nixos/lasuite-docs: add systemd service to create postgresql unaccent function
It is unsafe for a non-superadmin user to create a C functin
in postgresql, so we need to create the function via the
superadmin postgres user. This function (immutable unaccent) is
needed by lasuite-docs since 4.2.0 and is used to search for keys
without accents.

https://github.com/suitenumerique/docs/releases/tag/v4.2.0
https://github.com/suitenumerique/docs/compare/v4.1.0...v4.2.0
(cherry picked from commit cdfc90c4e5b56c615d40dc502f3cd9ef81e70d27)