nixos/opentabletdriver: fix "Daemon is already running"
This issue was observed after GNOME 48 -> 49 update.
User creation for gdm was changed, now it starts systemd user
services.
This causes issues for OTD and potentially other services, as
gdm-greeter user starts an OTD process creating a named pipe in /tmp
that acts as a lock file. When a user logs in, they also start a process
which exits with "Daemon is already running" and the gdm-greeter started
process exits.
Applying upstream patches that improve cleaning up the "lock file"
doesn't fix this issue, probably because there's not enough time between
stopping the first and starting the second process.

Fixes https://github.com/NixOS/nixpkgs/issues/469340.