nixos/test-driver: Use pytest assertion rewriting
This switches NixOS tests to use pytest's assertion rewriting mechanism,
so that we get more useful information from failures.

For example when forging a random test (adm-sev.nix) that uses
assertions to fail instead, without assertion rewriting the error looks
like this:

!!! Test "Check default settings" failed with error: ""

With assertion rewritiing it looks like this:

!!! Test "Check default settings" failed with error: "assert 'sev::' in 'root:x:0:\nwheel:x:1:\nkmem:x:2:\ntty:x:3:\nmessagebus:x:4:\ndisk:x:6:\naudio:x:17:\nfloppy:x:18:\nuucp:x:19:\nlp:x:2...xbld28,nixbld29,nixbld3,nixbld30,nixbld31,nixbld32,nixbld4,nixbld5,nixbld6,nixbld7,nixbld8,nixbld9\nnogroup:x:65534:\n'"

So you actually know what's going on and don't need to litter your code
with things like 'assert <foo>, "some context here"'.

I did this mainly because I like to write a lot of tests in my
deployments and do all the pytest assertion rewriting dance downstream,
but I thought it would be a good idea to actually have this available
for all NixOS tests.

The reason why I hesitated so far was that it relies on internal code of
pytest, but given that I already use this for the systemd-confinement
tests since almost two years and in my deployments for even longer
(around 4 years), I'd consider this to be a very minor maintenance
burden.

In addition to that, if we someday don't have or want to maintain this
part of the code anymore, we can simply drop it and all we get are less
useful assertions.

Signed-off-by: aszlig <aszlig@nix.build>