nixos/release: read configuration from usual places
Enable reading configuration from a module pointed by NIXOS_CONFIG
environment variable or <nixos-config> path as other modules do. If,
however, these are not set or set improperly an empty set is used by
default as before.

With this patch it slightly is more convenient to use nixos/release.nix
with nix-build for example to build customized netboot or iso images.