haskellPackages.MIP: fix broken, disable tests
The hackage zip is missing some of the test data,
thus the tests fail.
This was raised in https://github.com/msakai/haskell-MIP/issues/87
Until this is fixed upstream, we disable the tests.

With the fix of MIP, MIP-glpk is now broken not because of
MIP failing but because of a problem with the module itself.

The module doesn't build because of an incomatibility with
extended-reals >=0.2.7.0. This has been addressed upstream
https://github.com/msakai/haskell-MIP/commit/60a5ee234fc6667d2c34152761fa5a54a6f6e533
but has not been released, yet.