nixos/radicle: update behaviour of explicit systemd load creds options
We want to keep the default behaviour for basic options given as a plain
file path (without a colon) e.g. `whateverPasswordFile = "/my/path";`.
This should be interpreted as a path. But if a user opts-into explicit
systemd options they should be interpreted according to the normal
syntax of those options. In which case if there is no colon in the path,
like e.g. `whateverPasswordFile.LoadCredential = "my-cred";`, this
should not be interpreted as a path, but instead a credential ID.