python3.pkgs.scipy: use finalAttrs & make updates easier
Since these datasets haven't been changed for a while, and they are not
tied to a specific scipy version, it should be safe enough to put them
in the passthru, and make it easy to verify their hashes are up-to-date.

Also, using `buildConfig` Nix attribute set for the cross file .ini file
makes it easy to override it.