nixVersions.nix_2_28: unbreak builds on linux
Nix 2.28 seems to have been unbuildable for a long time due to
https://github.com/NixOS/nix/issues/11295.

This particularly surfaces in advanced CI setups where the store
locations are non standard and build user separation disappears as the
build system seems to be hiding bugs using these features.

Reproducing the issue that this commit fixes can be achieved by running:

$ nix-build '<nixpkgs>' -A nixVersions.nix_2_28 --store /tmp/alt --check

This only fixes Linux builds as they are the only "truly" affected by
this.

This commit can be reverted once upstream fixes this issue.

Change-Id: I55d7adee2dee7c735490f33395ba061c27cf4319
Signed-off-by: Raito Bezarius <raito@lix.systems>