git: add rust opt-in
With v2.52.0, git now includes the option to use a rust implementation
of some core logic [1]. Using rust/cargo as part of the build process is
currently still optional, though expected to become mandatory in Git
v3.0.

In preparation, introduce cargo/rustc into the Git package build. This
is disabled by default and enabled via a new `rustSupport` argument to
mirror the fact that this is opt-in in Git itself.

[1]: lore.kernel.org/git/20251002-b4-pks-rust-breaking-change-v8-0-3a89fd5b1ce7@pks.im/